requirements.txt
django
MySQL-python
enum34
south
Celery
boto
django-tastypie
pycrypto
requests
